paritial implementation of joml

+++ engine/src/main/java/org/terasology/rendering/world/ChunkMeshUpdateManager.java
@@ -17,11 +17,11 @@ package org.terasology.rendering.world;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Queues;
+import org.joml.Vector3f;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.terasology.math.ChunkMath;
-import org.terasology.math.geom.Vector3f;
-import org.terasology.math.geom.Vector3i;
+import org.joml.Vector3i;
 import org.terasology.monitoring.chunk.ChunkMonitor;
 import org.terasology.rendering.primitives.ChunkMesh;
 import org.terasology.rendering.primitives.ChunkTessellator;
@@ -195,7 +195,7 @@ public final class ChunkMeshUpdateManager {
             if (task.isTerminateSignal()) {
                 return -1;
             }
-            return distFromRegion(task.getPosition(), new Vector3i(cameraChunkPosX, cameraChunkPosY, cameraChunkPosZ));
+            return distFromRegion(task.getPosition(), new Vector3i((int) cameraChunkPosX, (int)cameraChunkPosY, (int)cameraChunkPosZ));
         }
 
         private int distFromRegion(Vector3i pos, Vector3i regionCenter) {
